error

+++ lcnn/models/line_vectorizer.py
@@ -331,7 +331,7 @@ class LineVectorizer(nn.Module):
             # )
             line = torch.cat([xyu[:, None], xyv[:, None]], 1)
             xy = xy.reshape(n_type, K, 2)
-            jcs = [xy[i, score[i]] for i in range(n_type)]
+            jcs = [xy[i, score[i].long()] for i in range(n_type)]
             # print("lines sample:", line.shape)
             # print("label", label.shape)
             return line, label, jcs
